Pymol-Inteface Update:  - cleaned Code  - converted docStrings to NumPY  - cleaning the test for import, export and optimizing
