+++ restraintmaker/algorithm/Selection.py
@@ -9,6 +9,8 @@ import restraintmaker.utils.Types
 import restraintmaker.utils.Utilities
 from restraintmaker.tools_Rdkit import Rdkit_Functions
 from restraintmaker.utils import Utilities as u
+
+
 # from restraintmaker.utils.Utilities import print
