+++ restraintmaker/interface_Pymol/RestraintMaker_Logic.py
@@ -1,10 +1,7 @@
 """
-.. automodule::  restraintmaker_Logic
-    :members:
-    TODO:write DOCU!
+    Contains the class Logic_Handler, which defines the program flow, decides which user activities are allowed and manages Importers, Selections,
+    Restraints, Filters, Optimzers and Exporters.
 
-Contains the class Logic_Handler, which defines the program flow, decides which user activities are allowed and manages Importers, Selections,
-Restraints, Filters, Optimzers and Exporters
 """
 import time
 import typing as t
@@ -17,50 +14,49 @@ from restraintmaker.algorithm import Filter
 from restraintmaker.algorithm import Optimizer
 from restraintmaker.algorithm import Selection
 from restraintmaker.interface_Pymol.Pymol_Utitlities import \
-    _convert_molecules_to_pdb  # TODO: Get independent of interface_Pymol +. wirte own conversion function
+    _convert_molecules_to_pdb, \
+    try_to_get_args  # TODO: Get independent of interface_Pymol +. wirte own conversion function
 from restraintmaker.io import Exporter
 from restraintmaker.io import Importer
 from restraintmaker.utils import Utilities as u, Types
-
-
-# Override print functions
-# Program verbosity can be set in utilities
-def print(*args, mv=0):
-    u.print(*args, mv=mv)
+from restraintmaker.utils.Utilities import print
 
 
 class Logic_Handler:
-    """class Logic_Handler: Handles the program flow. By defining it inside a class we get proper encapsulation and prevent problems with multiple imports
-
-        When a GUI Program Launches it should create one instance of a Logic_Handler and use it to control the program flow.
-        It should pass GUI-events to it (by calling the react_to_event method.). The GUI Module should NOT need to directly access any other modules. (Except utilities).
-        Warning: To keep things clean there is not backwards communicatopn from the Logic_Handler to the GUI module. It is the GUI Modules responsibility to check on the current state of the program, after creating events.
-        Warning: I really do NOT want the GUI-Program to inherit form the Logic_Handler. It will inherit from a GUI Element (e.g. interface_Pymol.Wizard). Python suppports multiple inheritance, but I don't.
-
-        RECOMMENDATIONS TO WRITE A NEW GUI MODULE:
-        The GUI Module should have the following functions/Elements
-
-        1) Buttons/Lists... for the different actions that can be performed (Import, RestraintType, Selection ...). These should call the function set_importer_type, set_restraint_type ...
-                The GUI Program should use the action_states dict, to check which actions are currently available.
-        2) 2 Toggle buttons / radio buttons... to switch between select / delete mode and atom / restraint mode
-
-        2) Events: The GUI Module is responsible for translating whatever events it understands into events relevant to the selections. It can do this by calling the function react_to_event(event_type, kw_args)
-           Recommended translations:    SELECT: When the user clicks a certain Atom
-                                        MOVE: Mouse movement / keyboard arrows ...
-                                        SIZE: Mouse wheel / keyboard arrows
-                                        CONFIRM: Double Click, Enter ...
-
-             The GUI module should check the state of the program (event_state dict and selected_atoms, selected-restraints) every time after calling react_to_event.
-             I recommend wrapping the react_to_event function into a function of the GUI module to do that automatically.
-
-        3) Some Selections (SphericalSelection & subclasses) are more intuitive if the GUI program draws a sphere at the corresponding position. => If you want that you have to introduce special for the GUI module to check the current selection
-    """
 
     def __init__(self, all_atoms: t.List[restraintmaker.utils.Utilities.Atom]):
-        '''
-        :param all_atoms: List of all atoms in the Molecules, which should be connected
-        :type all_atoms: t.List[u.Atom]
-        '''
+        """
+            Handles the program flow. By defining it inside a class we get proper encapsulation and prevent problems with multiple imports
+
+            When a GUI Program Launches it should create one instance of a Logic_Handler and use it to control the program flow.
+            It should pass GUI-events to it (by calling the react_to_event method.). The GUI Module should NOT need to directly access any other modules. (Except utilities).
+            Warning: To keep things clean there is not backwards communicatopn from the Logic_Handler to the GUI module. It is the GUI Modules responsibility to check on the current state of the program, after creating events.
+            Warning: I really do NOT want the GUI-Program to inherit form the Logic_Handler. It will inherit from a GUI Element (e.g. interface_Pymol.Wizard). Python suppports multiple inheritance, but I don't.
+
+            RECOMMENDATIONS TO WRITE A NEW GUI MODULE:
+            The GUI Module should have the following functions/Elements
+
+            1) Buttons/Lists... for the different actions that can be performed (Import, RestraintType, Selection ...). These should call the function set_importer_type, set_restraint_type ...
+                    The GUI Program should use the action_states dict, to check which actions are currently available.
+            2) 2 Toggle buttons / radio buttons... to switch between select / delete mode and atom / restraint mode
+
+            2) Events: The GUI Module is responsible for translating whatever events it understands into events relevant to the selections. It can do this by calling the function react_to_event(event_type, kw_args)
+               Recommended translations:    SELECT: When the user clicks a certain Atom
+                                            MOVE: Mouse movement / keyboard arrows ...
+                                            SIZE: Mouse wheel / keyboard arrows
+                                            CONFIRM: Double Click, Enter ...
+
+                 The GUI module should check the state of the program (event_state dict and selected_atoms, selected-restraints) every time after calling react_to_event.
+                 I recommend wrapping the react_to_event function into a function of the GUI module to do that automatically.
+
+            3) Some Selections (SphericalSelection & subclasses) are more intuitive if the GUI program draws a sphere at the corresponding position. => If you want that you have to introduce special for the GUI module to check the current selection
+
+        Parameters
+        ----------
+        all_atoms : t.List[u.Atom]
+             List of all atoms in the Molecules, which should be connected
+
+        """
 
         self.all_atoms: t.List[restraintmaker.utils.Utilities.Atom] = all_atoms
         self.selected_atoms: t.List[restraintmaker.utils.Utilities.Atom] = []
@@ -77,17 +73,17 @@ class Logic_Handler:
         self.my_importer = None
 
         # Define Selection-Types
-        self.available_selection_types: t.list[t.type] = u.get_all_subclasses(Selection._Selection)
+        self.available_selection_types: t.List[t.Type] = u.get_all_subclasses(Selection._Selection)
         self.current_selection_type = None
         self.my_selection = None  # Unlike my_filter my_selection needs to be accessed by several methods: do_pick, set_selection
 
         # Define Filter-Types
-        self.available_filter_types: t.list[t.type] = u.get_all_subclasses(Filter._Filter)
+        self.available_filter_types: t.List[t.Type] = u.get_all_subclasses(Filter._Filter)
         self.current_filter_type = None
         self.my_filter = None
 
         # Define-Optimizer Types
-        self.available_optimizer_types: t.List[t.type] = u.get_all_subclasses(Optimizer._Optimizer)
+        self.available_optimizer_types: t.List[t.Type] = u.get_all_subclasses(Optimizer._Optimizer)
         self.current_optimizer_type = None
         self.my_optimizer = None
 
@@ -97,7 +93,6 @@ class Logic_Handler:
         self.my_exporter = None
 
         # Modes & Action_States (Modes: What do to atoms, when they are selected. Action states: Which actions are possible right now)
-        # TODO CLEAN (1): If we ever use more porgram modes it mioght be cleaner to make the enum u.ActionStates a class,
         # which contains one attribute indicating actions AND one for each mode
         self.select_or_delete_mode = True  # True = select, False = delte
         self.atom_or_restraint_mode = True  # True = atom, False = Restraint
@@ -114,36 +109,42 @@ class Logic_Handler:
             'Done': restraintmaker.utils.program_states.ActionState.ALWAYS_ENABLED}
         self.set_action_states()
 
-    # END OF __init___
+    """
+        event_functions
+    """
 
-    # ----------------------------------------------------------------------------------------------------------------------------------------------------
-    # -----------------------------------------------------------------event_functions--------------------------------------------------------------------
-    # ----------------------------------------------------------------------------------------------------------------------------------------------------
     # These event functions will be called by the GUI-Program to indicate an action. It is the GUI program that defines what exactly consitutues what event. (e.g: a move event might be triggered by a mouse movement or keyboatrd arrows etc.depending on the GUI-Program)
     # CAREFULL: Never call the Slections _update*functions alone. Always check their state (finished or not & selected_atoms) directly afterwards in teh GUI program
 
     def react_to_event(self, event_type: restraintmaker.utils.program_states.EventType, **kw_args: t.Any):
-        '''
-        react_to_event should be called by the GUI-module to report an event
+        """
+                react_to_event should be called by the GUI-module to report an event
 
-        react_to_event will then call the relevant update function of the active selection and refresh all atom lists etc.
+            react_to_event will then call the relevant update function of the active selection and refresh all atom lists etc.
 
-        :warning: react_to_event will change the state of the program (i.e. selected_atoms, selected_restraints, mode and actions states)
-        It is strongly recommended to wrap react_to_event into a function of the GUI-module, which will check these and redraw the scene and de?activate buttons, if necessary/.
+         @Waring react_to_event will change the state of the program (i.e. selected_atoms, selected_restraints, mode and actions states)
+         It is strongly recommended to wrap react_to_event into a function of the GUI-module, which will check these and redraw the scene and de?activate buttons, if necessary/.
 
-        :param event_type: What kind of event was triggered?
-        :type event_type: restraintmaker.utils.program_states.EventType
-        :param kw_args: Arguments the relevant update function will need
-        :type kw_args: t.Any
-        :raises u.BadArgumentException if args to not match the the corresponding logic_handler.react_to_*_event_function
-        :raises NotImplementedError if a new event_type is used, which the function does not know about.
-        :return: None
-        :rtype: None
-        '''
 
-        # Note: Wierd python Syntax:  The * operator means two different things here: in the function definition **kw_args indicates, that I want to take an arbitrary number of keyword arguments and pack them all INTO a dict called args.
-        #                                                in the function calls ** kw_args means, that I want to UNPACK a dict called kwargs s and pass each value as a separate keyword-argument.
+        Parameters
+        ----------
+        event_type : restraintmaker.utils.program_states.EventType
+             What kind of event was triggered?
+        kw_args :  dict
+             Arguments the relevant update function will need
 
+        Returns
+        -------
+        NoReturn
+
+        Raises
+        ------
+        BadArgumentException
+            if args to not match the the corresponding logic_handler.react_to_*_event_function
+        NotImplementedError
+            if a new event_type is used, which the function does not know about.
+
+        """
         if self.my_selection == None or self.my_selection.has_finished:
             return
 
@@ -168,16 +169,28 @@ class Logic_Handler:
         self.set_action_states()
 
     def set_action_type(self, x_type):
-        '''set_action should be called by the GUI module when the user selects an Importer, Restraint, Selection, Filter, Optimizer or Exporter
+        """set_action should be called by the GUI module when the user selects an Importer, Restraint, Selection, Filter, Optimizer or Exporter
 
-        It will then call the relevant _set_X_type function.
+            It will then call the relevant _set_X_type function.
 
-        :warning: set_action_type will change the state of the program (i.e. selected_atoms, selected_restraints, mode and actions states)
-        It is strongly recommended to wrap react_to_event into a function of the GUI-module, which will check these and redraw the scene and de?activate buttons, if necessary/.
-        :param type: A type of Importer, Selection, Restraint, Filter, Optimizer or Exporter
-        :type type: type
-        :raise: TypeError if the specified type is not recognized
-        '''
+            @Warnings set_action_type will change the state of the program (i.e. selected_atoms, selected_restraints, mode and actions states)
+            It is strongly recommended to wrap react_to_event into a function of the GUI-module, which will check these and redraw the scene and de?activate buttons, if necessary/.
+
+        Parameters
+        ----------
+        x_type : type
+            A type of Importer, Selection, Restraint, Filter, Optimizer or Exporter
+
+        Returns
+        -------
+        NoReturn
+
+        Raises
+        ------
+        TypeError
+            if the specified type is not recognized
+
+        """
 
         # SWITCH-TABLE
         if x_type in self.available_importer_types:
@@ -200,21 +213,26 @@ class Logic_Handler:
         self._check_selection_status()
         self.set_action_states()
 
-    # TODO CLEAN (1): If we ever use more porgram modes it might be cleaner to combine action state and mode into one class
-    # and set_action_state and set_mode into one function.
     def set_mode(self, select_delete: bool, atom_restraint: bool):
-        '''
-        set_mode should be called by the GUI module when the user toggles the buttons to change the program mode
+        """
+            set_mode should be called by the GUI module when the user toggles the buttons to change the program mode
 
         warning: set_action_type will change the state of the program (i.e. selected_atoms, selected_restraints, mode and actions states)
         It is strongly recommended to wrap react_to_event into a function of the GUI-module, which will check these and redraw the scene and de?activate buttons, if necessary/.
-        :param select_delete: Select mode (True) or Delete mode (False)?
-        :type select_delete: bool
-        # :param atom_restraint: Atoms mode (True) or Delete mode (False)?
-        :type atom_restraint: bool
-        :return: None
-        :rtype: None
-        '''
+
+        Parameters
+        ----------
+        select_delete : bool
+            Select mode (True) or Delete mode (False)?
+        atom_restraint : bool
+            Atoms mode (True) or Delete mode (False)?
+
+        Returns
+        -------
+        NoReturn
+
+        """
+
         self.select_or_delete_mode = select_delete
         self.atom_or_restraint_mode = atom_restraint
 
@@ -239,12 +257,16 @@ class Logic_Handler:
 
     def _check_selection_status(self):
         """
-        _check_selection_status needs to be called at the end of every react_to_*_event function, after updating the selection.
+         _check_selection_status needs to be called at the end of every react_to_*_event function, after updating the selection.
 
          It checks if the selection has finished and adds the selected atoms to the list, if necessary.
-        :return: -
-        :rtype: -
+
+        Returns
+        -------
+        NoReturn
+
         """
+
         if self.my_selection == None:
             return
 
@@ -279,15 +301,12 @@ class Logic_Handler:
                 # open a new SingleAtomSelection
                 self.my_selection.reset()
 
-    # END OF _check_selection_status
+    """
+        set_*_type - functions:
+            Are called by set_action_type
+            used to be called do_import
+    """
 
-    # ----------------------------------------------------------------------------------------------------------------------------------------------------
-    # -----------------------------------------------------------------set_*_type()-----------------------------------------------------------------------
-    # ----------------------------------------------------------------------------------------------------------------------------------------------------
-
-    # Are called by set_action_type
-
-    # used to be called do_import
     def _set_importer_type(self, new_importer_type: t.Type[Importer._Importer]):
         '''
         _set_importer_type changes the current Importer type, and creates and apllies a new Importer of that type.
@@ -346,15 +365,21 @@ class Logic_Handler:
         self._set_selection_type(self.current_restraint_type.accepted_selection_types[0])
 
     def _set_selection_type(self, new_selection_type: type, ):
-        '''
-        _set_selection_type chagnes the current type of Selection and starts a new selection of that type.
+        """
+             chagnes the current type of Selection and starts a new selection of that type.
 
-        :warning: Should only be called via set_action_type which handles type checking
-        :param new_selection_type: type of the new Selection
-        :type new_selection_type: t.Type[Selection]
-        :return: None
-        :rtype: None
-        '''
+            @warnings: Should only be called via set_action_type which handles type checking
+
+        Parameters
+        ----------
+        new_selection_type : t.Type[Selection]
+            type of the new Selection
+
+        Returns
+        -------
+        NoReturn
+
+        """
 
         if new_selection_type == None:
             self.current_selection_type = None
@@ -384,17 +409,22 @@ class Logic_Handler:
 
         # Unlike other classes Selection does not call its own method (update(...)) here. It is just started. update(...) is called at every pick
 
-    # Actually: Also consider just keeping it simple, as it is
     def _set_filter_type(self, new_filter_type: type):
-        '''
-        _set_filter_type changes the current type of Filter and creates and applies a Filter of that type.
+        """
+            changes the current type of Filter and creates and applies a Filter of that type.
 
-        :warning: Should only be called via set_action_type which handles type checking
-        :param new_filter_type: type of the new Filter
-        :type new_filter_type: t.Type[Filter]
-        :return: None
-        :rtype: None
-        '''
+            @Warnings should only be called via set_action_type which handles type checking
+
+        Parameters
+        ----------
+        new_filter_type :  t.Type[Filter]
+            type of the new Filter
+
+        Returns
+        -------
+        NoReturn
+
+        """
 
         if new_filter_type == None:
             self.current_filter_type = None
@@ -414,9 +444,11 @@ class Logic_Handler:
 
         input_function = u.do_nothing
         if isinstance(self.my_filter, Filter.PropertyFilter):
-            input_function = restraintmaker.utils.qt_dialogs.create_input_dialog(parent_window=None, title='PropertyFilter')
+            input_function = restraintmaker.utils.qt_dialogs.create_input_dialog(parent_window=None,
+                                                                                 title='PropertyFilter')
         elif isinstance(self.my_filter, Filter.ElementFilter):
-            input_function = restraintmaker.utils.qt_dialogs.create_input_dialog(parent_window=None, title='ElementFilter')
+            input_function = restraintmaker.utils.qt_dialogs.create_input_dialog(parent_window=None,
+                                                                                 title='ElementFilter')
         elif isinstance(self.my_filter, Filter.RingFilter):
             # TODO: GET INDEPENDENT OF PYMOL HERE => Write an own pdb function
             input_function = lambda dummy_arg: _convert_molecules_to_pdb()
@@ -428,15 +460,21 @@ class Logic_Handler:
             self.current_filter_type = None
 
     def _set_optimizer_type(self, new_optimizer_type: type):
-        '''
-        _set_optimizer_type changes the current type of Optimizer and creates and applies a Optimizer of that type.
+        """
+            changes the current type of Optimizer and creates and applies a Optimizer of that type.
+            @Warnings  Should only be called via set_action_type which handles type checking.
 
-        :warning: Should only be called via set_action_type which handles type checking.
-        :param new_optimizer_type: type of the new Filter
-        :type new_optimizer_type: t.Type[Filter]
-        :return: None
-        :rtype: None
-        '''
+        Parameters
+        ----------
+        new_optimizer_type : t.Type[Filter]
+            type of the new Filter
+
+        Returns
+        -------
+        NoReturn
+
+        """
+
         if new_optimizer_type == None:
             self.current_optimizer_type = None
             self.my_optimizer = None
@@ -447,46 +485,46 @@ class Logic_Handler:
 
         self.my_optimizer = new_optimizer_type(self.selected_atoms)
 
-        # TODO: See TODO in u.creat_multi_dialog
         # SWITCHTABLE FOR input function
         input_functions = u.do_nothing()
-        # TODO: More elegant than these hardcoded input functions: Give each Optimizer etc. an attribute specifzing which args it needs, which we can than pass to create_multi_dialog
-        # TODO: Generate new field in pyqt win to select if you want a chain, ring, or all to all connection
 
         if isinstance(self.my_optimizer, Optimizer.TreeHeuristicOptimizer):
-            input_function = restraintmaker.utils.qt_dialogs.create_multi_dialog(title='Parameters for TreeHeuristicOptimizer', \
-                                                                                 inputs=['number of restraints', 'maximal distance of restraints',
-                                                           'tree-algorithm', 'optimize molecules pairs by'], \
-                                                                                 options={'tree-algorithm': ['shortest', 'cog', 'prim', "biased_avg"],
-                                                            'optimize molecules pairs by': ['None', 'convex_hull',
-                                                                                            'pca_2d']}, \
-                                                                                 default={'number of restraints': '4',
-                                                            'maximal distance of restraints': '1.2',
-                                                            'algorithm': 'shortest',
-                                                            'optimize molecules pairs by': 'pca_2d'})
+            input_function = restraintmaker.utils.qt_dialogs.create_multi_dialog(
+                title='Parameters for TreeHeuristicOptimizer', \
+                inputs=['number of restraints', 'maximal distance of restraints',
+                        'tree-algorithm', 'optimize molecules pairs by'], \
+                options={'tree-algorithm': ['shortest', 'cog', 'prim', "biased_avg"],
+                         'optimize molecules pairs by': ['None', 'convex_hull',
+                                                         'pca_2d']}, \
+                default={'number of restraints': '4',
+                         'maximal distance of restraints': '1.2',
+                         'algorithm': 'shortest',
+                         'optimize molecules pairs by': 'pca_2d'})
 
 
         elif isinstance(self.my_optimizer, Optimizer.BruteForceRingOptimzer):
-            input_function = restraintmaker.utils.qt_dialogs.create_multi_dialog(title='Parameters for BruteForceOptimizer', \
-                                                                                 inputs=['number of restraints', 'maximal distance of restraints',
-                                                           'algorithm', 'optimize molecules pairs by'], \
-                                                                                 options={'algorithm': ['convex_hull', 'pca'],
-                                                            'optimize molecules pairs by': ['None', 'convex_hull',
-                                                                                            'pca_2d']}, \
-                                                                                 default={'number of restraints': '4',
-                                                            'maximal distance of restraints': '1.2', 'algorithm': 'pca',
-                                                            'optimize molecules pairs by': 'pca_2d'})
+            input_function = restraintmaker.utils.qt_dialogs.create_multi_dialog(
+                title='Parameters for BruteForceOptimizer', \
+                inputs=['number of restraints', 'maximal distance of restraints',
+                        'algorithm', 'optimize molecules pairs by'], \
+                options={'algorithm': ['convex_hull', 'pca'],
+                         'optimize molecules pairs by': ['None', 'convex_hull',
+                                                         'pca_2d']}, \
+                default={'number of restraints': '4',
+                         'maximal distance of restraints': '1.2', 'algorithm': 'pca',
+                         'optimize molecules pairs by': 'pca_2d'})
 
         elif isinstance(self.my_optimizer, Optimizer.MetaMoleculeRingOptimizer):
-            input_function = restraintmaker.utils.qt_dialogs.create_multi_dialog(title='Parameters for BestMoleculeRingOptimizer', \
-                                                                                 inputs=['number of restraints', 'maximal distance of restraints',
-                                                           'algorithm', 'optimize molecules pairs by'], \
-                                                                                 options={'algorithm': ['convex_hull', 'pca'],
-                                                            'optimize molecules pairs by': ['convex_hull',
-                                                                                            'pca_2d']}, \
-                                                                                 default={'number of restraints': '4',
-                                                            'maximal distance of restraints': '1.2', 'algorithm': 'pca',
-                                                            'optimize molecules pairs by': 'pca_2d'})
+            input_function = restraintmaker.utils.qt_dialogs.create_multi_dialog(
+                title='Parameters for BestMoleculeRingOptimizer', \
+                inputs=['number of restraints', 'maximal distance of restraints',
+                        'algorithm', 'optimize molecules pairs by'], \
+                options={'algorithm': ['convex_hull', 'pca'],
+                         'optimize molecules pairs by': ['convex_hull',
+                                                         'pca_2d']}, \
+                default={'number of restraints': '4',
+                         'maximal distance of restraints': '1.2', 'algorithm': 'pca',
+                         'optimize molecules pairs by': 'pca_2d'})
 
         if try_to_get_args(self.my_optimizer, input_function):
             time_start = time.time()
@@ -502,15 +540,20 @@ class Logic_Handler:
             self.current_optimizer_type = None
 
     def _set_exporter_type(self, new_exporter_type: type):
-        '''
-            _set_exporter_type changes the current type of Exporter and creates and applies an Exporter of that type.
+        """
+            changes the current type of Exporter and creates and applies an Exporter of that type.
+            @Warnings  Should only be called via set_action_type which handles type checking.
 
-            :warning: Should only be called via set_action_type which handles type checking.
-            :param new_exporter_type: type of the new Filter
-            :type new_exporter_type: t.Type[Exporter]
-            :return: None
-            :rtype: None
-            '''
+        Parameters
+        ----------
+        new_exporter_type :  t.Type[Exporter]
+            type of the new Filter
+
+        Returns
+        -------
+        NoReturn
+
+        """
 
         if new_exporter_type == None:
             self.current_exporter_type == None
@@ -521,69 +564,48 @@ class Logic_Handler:
             raise TypeError(str(new_exporter_type) + ' is not an available type of Exporter')
 
         self.my_exporter = new_exporter_type(self.selected_restraints)
-        if try_to_get_args(self.my_exporter, restraintmaker.utils.qt_dialogs.create_file_save_dialog()):  # u.create_input_dialog(None, 'Exporter')
+        if try_to_get_args(self.my_exporter,
+                           restraintmaker.utils.qt_dialogs.create_file_save_dialog()):  # u.create_input_dialog(None, 'Exporter')
             self.my_exporter.export_restraints()
         else:
             self.my_exporter = None
             self.current_exporter_type = None
 
     def set_action_states(self):
-        """set_action_statess checks the current state of the program and decides, which actions are allowed at the moment"""
+        """
+            checks the current state of the program and decides, which actions are allowed at the moment
+
+            many TODOs
+            
+        Returns
+        -------
+        NoReturn
+
+        """
 
         # ALL CHANGES TO THE ACTION STATE SHOULD BE DONE IN THIS FUNCTION. THE STATES SHOULD BE DETERMINED ONLY BY THE CURRENT STAte of THE PROGRAM
         # ORDER: Set types for each action one after the other
         # if the structure beomces more complex consider first checking the toggle modes (atom?restraint, delete/select) and go through each state in iach state
-
         self._action_states['toggle_select_delete']: restraintmaker.utils.program_states.ActionState.ALWAYS_ENABLED
         self._action_states['toggle_atom_restraint']: restraintmaker.utils.program_states.ActionState.ALWAYS_ENABLED
         self._action_states['Importer'] = restraintmaker.utils.program_states.ActionState.CURRENTLY_ENABLED if len(
             self.selected_restraints) == 0 else restraintmaker.utils.program_states.ActionState.CURRENTLY_DISABLED  # TODO Later: Check if there is any Molecules loaded.
         self._action_states['Restraint'] = restraintmaker.utils.program_states.ActionState.CURRENTLY_ENABLED if (
-                                                                                  not self.atom_or_restraint_mode) and self.select_or_delete_mode \
+                                                                                                                    not self.atom_or_restraint_mode) and self.select_or_delete_mode \
             else restraintmaker.utils.program_states.ActionState.CURRENTLY_DISABLED  # TODO Later: Always Enabled, can be changed as much as we like
         self._action_states[
             'Selection'] = restraintmaker.utils.program_states.ActionState.CURRENTLY_ENABLED if self.atom_or_restraint_mode else restraintmaker.utils.program_states.ActionState.CURRENTLY_DISABLED  # TODO Later: Check if any atoms are loaded
-        self._action_states['Filter'] = restraintmaker.utils.program_states.ActionState.CURRENTLY_ENABLED if len(self.selected_atoms) > 0 \
+        self._action_states['Filter'] = restraintmaker.utils.program_states.ActionState.CURRENTLY_ENABLED if len(
+            self.selected_atoms) > 0 \
                                                                                                              and len(
             self.selected_restraints) == 0 else restraintmaker.utils.program_states.ActionState.CURRENTLY_DISABLED
-        self._action_states['Optimizer'] = restraintmaker.utils.program_states.ActionState.CURRENTLY_ENABLED if len(self.selected_atoms) >= 2 \
+        self._action_states['Optimizer'] = restraintmaker.utils.program_states.ActionState.CURRENTLY_ENABLED if len(
+            self.selected_atoms) >= 2 \
                                                                                                                 and len(
             self.selected_restraints) == 0 else restraintmaker.utils.program_states.ActionState.CURRENTLY_DISABLED
         # TODO: Check if we have Atoms from at least 2 Molecules
         # TODO: Check indivudiual Optimizers depending on Restraint type
-        self._action_states['Exporter'] = restraintmaker.utils.program_states.ActionState.CURRENTLY_ENABLED if len(self.selected_restraints) > 0 \
+        self._action_states['Exporter'] = restraintmaker.utils.program_states.ActionState.CURRENTLY_ENABLED if len(
+            self.selected_restraints) > 0 \
             else restraintmaker.utils.program_states.ActionState.CURRENTLY_DISABLED
         self._action_states['Done'] = restraintmaker.utils.program_states.ActionState.ALWAYS_ENABLED
-
-
-# ------------------END OF CLASS------------------------------------------------------
-
-
-def try_to_get_args(my_x, input_function) -> bool:
-    '''
-    wrapper function for get_args. Takes care of the error handling
-
-    :param my_x: An Object that should call get args
-    :type my_x: _Importer,_Restraint,_Selection,_Filter,_Optimzer or _Exporter
-    :param input_function: The input_function argument of get_args
-    :type input_function: t.Callable[str]
-    :return: Could get_args be executed without errors?
-    :rtype: bool
-    '''
-
-    failed = False
-
-    try:
-        my_x.get_args(input_function)
-    except restraintmaker.utils.Utilities.BadArgumentException as err:
-        print('Failed to create a ' + my_x.__class__.__name__ + ' with the given arguments: \n' + str(err), mv=4)
-        failed = True
-    except IndexError:
-        print('Did not receive enough arguments to create a ' + my_x.__class__.__name__, mv=4)
-        # TODO: Would be nice to know how many args are needed here. See also TODO before u.create_multi_dialog
-        failed = True
-    except NotImplementedError:
-        print(my_x.__class__.__name__ + 'has not been implemented', mv=4)
-        failed = True
-    # Do not expect Argumetn error that will ocurr if my_x does not have a get_args function. That is a fault that really should stop the program . => Leave it to standard error handling
-    return not failed
