+++ restraintmaker/io/Files/Gromos_files.py
@@ -3,7 +3,6 @@ THIS CODE WAS BORROWED FROM PYGROMOSTOOLS - Check it out :)
 
 """
 
-
 # imports
 from typing import *
