+++ restraintmaker/test/test_files/IO/in_5ligs_disres.dat
@@ -1,0 +1,50 @@
+TITLE
+generated disres file for BRD4 with PYMOL wizard
+	 >>> Generated with python lib function_libs utilities. (riniker group)
+	 >>> line_seperator: ' \n'	 field_seperator: ' \t '
+END
+DISTANCERESSPEC 
+# KDISH 	 KDISC 
+ 	 0.1 	 0.153 
+#   i     j     k     l  type        i     j     k     l  type          r0         w0 rah
+##	KL1/H11 116 - KL20/H11 156
+  116     0     0     0     0      156     0     0     0     0     0.00300    1.00000   1
+##	KL1/H28 120 - KL20/H28 160
+  120     0     0     0     0      160     0     0     0     0     0.00500    1.00000   1
+##	KL1/HN11 111 - KL20/HN11 151
+  111     0     0     0     0      151     0     0     0     0     0.00300    1.00000   1
+##	KL1/H20 118 - KL20/H20 158
+  118     0     0     0     0      158     0     0     0     0     0.00200    1.00000   1
+##	KL17/H11 34 - KL20/H11 156
+   34     0     0     0     0      156     0     0     0     0     0.00100    1.00000   1
+##	KL17/H23 38 - KL20/H28 160
+   38     0     0     0     0      160     0     0     0     0     0.00500    1.00000   1
+##	KL17/HN11 29 - KL20/HN11 151
+   29     0     0     0     0      151     0     0     0     0     0.00300    1.00000   1
+##	KL17/H20 36 - KL20/H20 158
+   36     0     0     0     0      158     0     0     0     0     0.00300    1.00000   1
+##	KL19/H11 76 - KL1/H11 116
+   76     0     0     0     0      116     0     0     0     0     0.00400    1.00000   1
+##	KL19/H29 80 - KL1/H28 120
+   80     0     0     0     0      120     0     0     0     0     0.02100    1.00000   1
+##	KL19/HN11 71 - KL1/HN11 111
+   71     0     0     0     0      111     0     0     0     0     0.00300    1.00000   1
+##	KL19/H20 78 - KL1/H20 118
+   78     0     0     0     0      118     0     0     0     0     0.00700    1.00000   1
+##	KL17/H11 34 - KL21/H11 197
+   34     0     0     0     0      197     0     0     0     0     0.00100    1.00000   1
+##	KL17/H23 38 - KL21/H30 201
+   38     0     0     0     0      201     0     0     0     0     0.01500    1.00000   1
+##	KL17/HN11 29 - KL21/HN11 192
+   29     0     0     0     0      192     0     0     0     0     0.00400    1.00000   1
+##	KL17/H20 36 - KL21/H20 199
+   36     0     0     0     0      199     0     0     0     0     0.00100    1.00000   1
+##	KL19/H11 76 - KL21/H11 197
+   76     0     0     0     0      197     0     0     0     0     0.00100    1.00000   1
+##	KL19/H29 80 - KL21/H30 201
+   80     0     0     0     0      201     0     0     0     0     0.00900    1.00000   1
+##	KL19/HN11 71 - KL21/HN11 192
+   71     0     0     0     0      192     0     0     0     0     0.00100    1.00000   1
+##	KL19/H20 78 - KL21/H20 199
+   78     0     0     0     0      199     0     0     0     0     0.00400    1.00000   1
+END
