--- restraintmaker/test/test_files/IO/out_disres.dat
@@ -1,52 +1,0 @@
-TITLE
-generated disres file for BRD4 with PYMOL wizard
-	 >>> Generated with python lib function_libs utilities. (riniker group)
-	 >>> line_seperator: ' \n'	 field_seperator: ' \t '
-END
-DISTANCERESSPEC 
-# KDISH 	 KDISC 
- 	 0.1 	 0.153 
-#   i     j     k     l  type        i     j     k     l  type          r0         w0 rah
-##	3MXF/C18 2 - 4J3I/C27 62
-    2     0     0     0     0       62     0     0     0     0     0.08800    1.00000   1
-##	3MXF/C32 35 - 4J3I/C10 46
-   35     0     0     0     0       46     0     0     0     0     0.05400    1.00000   1
-##	3MXF/C31 29 - 4J3I/C14 36
-   29     0     0     0     0       36     0     0     0     0     0.11600    1.00000   1
-##	4J3I/O1 48 - 4MR3/O20 97
-   48     0     0     0     0       97     0     0     0     0     0.07900    1.00000   1
-##	4J3I/C14 36 - 4MR3/C6 82
-   36     0     0     0     0       82     0     0     0     0     0.06000    1.00000   1
-##	4J3I/C10 46 - 4MR3/C26 94
-   46     0     0     0     0       94     0     0     0     0     0.05800    1.00000   1
-##	4MR3/O1 81 - 4E96/O2 112
-   81     0     0     0     0      112     0     0     0     0     0.07800    1.00000   1
-##	4MR3/H21 98 - 4E96/O11 128
-   98     0     0     0     0      128     0     0     0     0     0.09500    1.00000   1
-##	4MR3/C26 94 - 4E96/C14 131
-   94     0     0     0     0      131     0     0     0     0     0.08200    1.00000   1
-##	4E96/C14 131 - 4LYS/O16 164
-  131     0     0     0     0      164     0     0     0     0     0.11300    1.00000   1
-##	4E96/O19 100 - 4LYS/H48 142
-  100     0     0     0     0      142     0     0     0     0     0.10000    1.00000   1
-##	4E96/H27 126 - 4LYS/C14 159
-  126     0     0     0     0      159     0     0     0     0     0.10700    1.00000   1
-##	4LYS/C7 152 - 4LYW/C1 177
-  152     0     0     0     0      177     0     0     0     0     0.11500    1.00000   1
-##	4LYS/C17 165 - 4LYW/C26 171
-  165     0     0     0     0      171     0     0     0     0     0.08800    1.00000   1
-##	4LYS/C20 139 - 4LYW/N7 180
-  139     0     0     0     0      180     0     0     0     0     0.08300    1.00000   1
-##	4LYW/C9 183 - 3U5L/Cl1 216
-  183     0     0     0     0      216     0     0     0     0     0.06000    1.00000   1
-##	4LYW/O30 168 - 3U5L/N9 203
-  168     0     0     0     0      203     0     0     0     0     0.09700    1.00000   1
-##	4LYW/O30 168 - 3U5L/N8 204
-  168     0     0     0     0      204     0     0     0     0     0.09700    1.00000   1
-##	3U5L/C11 201 - 3MXF/C32 35
-  201     0     0     0     0       35     0     0     0     0     0.03100    1.00000   1
-##	3U5L/H30 227 - 3MXF/Cl1 20
-  227     0     0     0     0       20     0     0     0     0     0.08400    1.00000   1
-##	3U5L/C21 210 - 3MXF/C12 8
-  210     0     0     0     0        8     0     0     0     0     0.05600    1.00000   1
-END
